Minor changes and moving things around
